Fix remaining Groovy sandbox HTTP/thread violations and expand run-all reporting. Migrate CrafterQ, Slack, MCP, fetch/post/preview helpers, and orchestration LLM/image paths off HttpURLConnection to StudioAiSandboxHttp; add ParallelToolThreadFactory so TranslateContentBatch can spawn pool threads under the Studio sandbox. Harden proxyImage (image/* only, non-empty body, nosniff) and restore pipeline heartbeat hints for RestClient completion phases. Extend the integration harness with per-test JSONL reporting, full tool/recipe matrix scenarios, concurrent session checks, and partial-failure handling for optional integration tools.

+++ authoring/scripts/classes/plugins/org/craftercms/aiassistant/studio/contrib/tool/builtin/integrations/WebSearchTool.groovy
@@ -111,7 +111,7 @@ class WebSearchTool extends AbstractStudioAiTool {
    */
   private static int maxResults(Integer toolRequested) {
     int r = (toolRequested != null) ? toolRequested.intValue() : DEFAULT_MAX_RESULTS
-    return Math.min(15, Math.max(1, r))
+    return (int) Math.min(15, Math.max(1, r))
   }
 
   /**
